Fixes #665

+++ inst/sql/sql_server/export/condition/sqlConditionsByType.sql
@@ -1,5 +1,5 @@
 with summary as (
-  select cast(stratum_1 as int) condition_concept_id, cast(stratum_2 as int) condition_type_concept_id, count_value, sum(count_value) s_count_value
+  select cast(stratum_1 as bigint) condition_concept_id, cast(stratum_2 as bigint) condition_type_concept_id, count_value, sum(count_value) s_count_value
   FROM  @results_database_schema.achilles_results
   where analysis_id = 405
   GROUP BY stratum_1, stratum_2, count_value
