+++ inst/sql/sql_server/export/device/sqlDevicesByType.sql
@@ -4,7 +4,7 @@ select c1.concept_id as OBSERVATION_CONCEPT_ID,
 	c2.concept_name as CONCEPT_NAME, 
 	ar1.count_value as COUNT_VALUE
 from (
-  select cast(stratum_1 as int) stratum_1, cast(stratum_2 as int) stratum_2, count_value
+  select cast(stratum_1 as bigint) stratum_1, cast(stratum_2 as bigint) stratum_2, count_value
   from @results_database_schema.achilles_results
   where analysis_id = 2105
   GROUP BY analysis_id, stratum_1, stratum_2, count_value
