+++ inst/sql/sql_server/export/measurement/sqlMeasurementTable.sql
@@ -5,12 +5,12 @@ select concept.concept_id,
 	round(1.0*ar2.count_value / ar1.count_value,5) as records_per_person,
 	ar3.stratum_3 as percent_missing_values
 from 
-	(select cast(stratum_1 as int) stratum_1, count_value from @results_database_schema.achilles_results where analysis_id = 1800 GROUP BY analysis_id, stratum_1, count_value) ar1
+	(select cast(stratum_1 as bigint) stratum_1, count_value from @results_database_schema.achilles_results where analysis_id = 1800 GROUP BY analysis_id, stratum_1, count_value) ar1
 	inner join
-	(select cast(stratum_1 as int) stratum_1, count_value from @results_database_schema.achilles_results where analysis_id = 1801 GROUP BY analysis_id, stratum_1, count_value) ar2
+	(select cast(stratum_1 as bigint) stratum_1, count_value from @results_database_schema.achilles_results where analysis_id = 1801 GROUP BY analysis_id, stratum_1, count_value) ar2
 	on ar1.stratum_1 = ar2.stratum_1
 	left join
-	(select cast(stratum_1 as int) stratum_1, round(cast(stratum_3 as float),4) as stratum_3 from @results_database_schema.achilles_results where analysis_id = 1833) ar3
+	(select cast(stratum_1 as bigint) stratum_1, round(cast(stratum_3 as float),4) as stratum_3 from @results_database_schema.achilles_results where analysis_id = 1833) ar3
 	on ar2.stratum_1 = ar3.stratum_1
 	inner join
 	(
