+++ inst/sql/sql_server/export/observationperiod/ageatfirst.sql
@@ -1,4 +1,4 @@
-select cast(ar1.stratum_1 as int) as interval_index, 
+select cast(ar1.stratum_1 as bigint) as interval_index, 
 	ar1.count_value as count_value, 
 	round(1.0*ar1.count_value / denom.count_value,5) as percent_value
 from
