+++ inst/sql/sql_server/export/observationperiod/observedbyyear_data.sql
@@ -1,4 +1,4 @@
-select cast(ar1.stratum_1 as int) - MinValue.MinValue as interval_index, 
+select cast(ar1.stratum_1 as bigint) - MinValue.MinValue as interval_index, 
   ar1.count_value as count_value, 
   round(1.0*ar1.count_value / denom.count_value,5) as percent_value
 from 
@@ -6,7 +6,7 @@ from
 	select * from @results_database_schema.achilles_results where analysis_id = 109
 ) ar1,
 (
-	select min(cast(stratum_1 as int)) as MinValue 
+	select min(cast(stratum_1 as bigint)) as MinValue 
 	from @results_database_schema.achilles_results where analysis_id = 109
 ) MinValue,
 (
