+++ inst/sql/sql_server/export/person/population_age_gender.sql
@@ -3,8 +3,8 @@ select c1.concept_id as concept_id,
   	ar1.stratum_2 as age,
 	  ar1.count_value as count_value
 from (
-  select cast(stratum_1 as int) stratum_1,
-	cast(stratum_2 as int) stratum_2, 
+  select cast(stratum_1 as bigint) stratum_1,
+	cast(stratum_2 as bigint) stratum_2, 
 	count_value
   from @results_database_schema.achilles_results
   where analysis_id = 102
